moved puzzle piece to it's own file.
